+++ CHANGELOG.md
@@ -7,9 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- The `Collection` classes now extends `Array`, so they are proper array objects.
-- Added the utility methods `collection()`, `successCollection()` and
-  `failureCollection()` for creating instances of the corresponding classes.
+- Removed all the `collection` classes. The `all()` and `allSettled()` methods
+  instead return `Result` implementations.
 
 ## [0.0.3-beta.0] - 2020-05-01